Add files via upload
Version 3 experienced some issues with the "Arachne" wall generator, which led to print failures when using Anycubic SlicerNext. In version 3.1, the wall generator has been switched to "Classic" in all SlicerNext profiles. This is the only significant change! The only other modification is an additional wipe in the start code.

In the Orca profiles, the wall generator remains set to "Arachne" because it produces better results for thin walls, and the Orca slicer does not encounter any issues with Arachne.